adding two insert function, one =with range and one with count values, working on insert with initializer list

+++ srcs/vector/Vector_iterator.hpp
@@ -41,12 +41,12 @@ class Vector_iterator
 			return (base.ptr == this->ptr) ;
 		}
 
-		Vector_iterator<T> operator+(int i) {
+		Vector_iterator<T> operator+(size_t i) {
 			Vector_iterator<T> tmp(ptr + i);
 			return (tmp);
 		}
 
-		Vector_iterator<T> operator-(int i) {
+		Vector_iterator<T> operator-(size_t i) {
 			Vector_iterator<T> tmp(ptr - i);
 			return (tmp);
 		}
